requirements.txt
fastapi
uvicorn
gunicorn
pycryptodome
colored
